Day 3
I had a subtle bug in part two due to a pair of simplifications,
which caused me to miss a pair of 9's if there was a subsequent 9
after them.

My son sent me his input and the list of his intermediate values,
which allowed me to run my code over his input and see where we
differed (first on line 75).  This saved me a ton of time over
finding the failing test case.